Split architecture and supported features
