+++ internal/clients/http/ec2/aws_supported_types.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/RHEnVision/provisioning-backend/internal/clients"
+	"github.com/RHEnVision/provisioning-backend/internal/clients/supported"
 	"github.com/RHEnVision/provisioning-backend/internal/payloads"
 	"github.com/aws/aws-sdk-go-v2/service/ec2/types"
 )
@@ -24,7 +25,7 @@ func NewInstanceTypes(ctx context.Context, types []types.InstanceTypeInfo) ([]*c
 				Cores:        *types[i].VCpuInfo.DefaultCores,
 				MemoryMiB:    *types[i].MemoryInfo.SizeInMiB,
 				Architecture: arch,
-				Supported:    clients.IsSupported(string(types[i].InstanceType)),
+				Supported:    supported.IsSupported(string(types[i].InstanceType)),
 			}
 			if types[i].InstanceStorageInfo != nil {
 				it.EphemeralStorageGB = *types[i].InstanceStorageInfo.TotalSizeInGB
