+++ internal/clients/supported/package.go
@@ -1,0 +1,3 @@
+// Functionality from this package will be moved into cmd/typesctl command.
+// TODO: Move and delete this package.
+package supported
